[#36] refactor: 패키지 구조 리팩토링
application layer 추가
