+++ src/main/java/dabang/star/cafe/api/CurrentMemberApi.java
@@ -6,9 +6,9 @@ import dabang.star.cafe.api.aop.SessionId;
 import dabang.star.cafe.api.request.CurrentMemberRequest;
 import dabang.star.cafe.api.request.MemberUpdateRequest;
 import dabang.star.cafe.domain.member.MemberData;
-import dabang.star.cafe.domain.login.MemberLoginService;
+import dabang.star.cafe.domain.authentication.MemberLoginService;
 import dabang.star.cafe.domain.member.Member;
-import dabang.star.cafe.domain.member.MemberService;
+import dabang.star.cafe.application.MemberService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
