+++ src/main/java/dabang/star/cafe/api/ManagerApi.java
@@ -2,7 +2,7 @@ package dabang.star.cafe.api;
 
 import dabang.star.cafe.api.request.ManagerLoginRequest;
 import dabang.star.cafe.domain.manager.ManagerData;
-import dabang.star.cafe.domain.login.ManagerLoginService;
+import dabang.star.cafe.domain.authentication.ManagerLoginService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
