+++ src/main/java/dabang/star/cafe/application/MemberService.java
@@ -1,21 +1,21 @@
-package dabang.star.cafe.domain.member;
+package dabang.star.cafe.application;
 
 import dabang.star.cafe.api.exception.DuplicatedException;
 import dabang.star.cafe.api.exception.MemberNotFoundException;
-import dabang.star.cafe.domain.login.EncryptService;
+import dabang.star.cafe.domain.authentication.EncryptService;
+import dabang.star.cafe.domain.member.Member;
+import dabang.star.cafe.domain.member.MemberData;
+import dabang.star.cafe.domain.member.MemberRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
 
-@Transactional
 @RequiredArgsConstructor
 @Service
-public class MemberServiceImpl implements MemberService {
+public class MemberService {
 
     private final MemberRepository memberRepository;
     private final EncryptService encryptService;
 
-    @Override
     public MemberData join(String email, String password, String nickname, String telephone, String birth) {
 
         checkDuplicatedEmail(email);
@@ -33,7 +33,6 @@ public class MemberServiceImpl implements MemberService {
         return MemberData.from(member);
     }
 
-    @Override
     public void checkDuplicatedEmail(String email) {
 
         if (memberRepository.existsByEmail(email)) {
@@ -41,20 +40,17 @@ public class MemberServiceImpl implements MemberService {
         }
     }
 
-    @Override
     public void update(Member member) {
         member.setPassword(encryptService.encrypt(member.getPassword()));
         memberRepository.save(member);
     }
 
-    @Override
     public MemberData loadById(long id) {
         return memberRepository.findById(id).orElseThrow(
                 () -> new MemberNotFoundException("member not found")
         );
     }
 
-    @Override
     public void secession(long id) {
         memberRepository.deleteById(id);
     }
