+++ src/main/java/dabang/star/cafe/domain/authentication/EncryptService.java
@@ -1,4 +1,4 @@
-package dabang.star.cafe.domain.login;
+package dabang.star.cafe.domain.authentication;
 
 public interface EncryptService {
