+++ src/main/java/dabang/star/cafe/domain/authentication/MemberLoginService.java
@@ -1,4 +1,4 @@
-package dabang.star.cafe.domain.login;
+package dabang.star.cafe.domain.authentication;
 
 
 import dabang.star.cafe.domain.member.MemberData;
